fixed z-stacking for bleed images
