+++ components/blocks/imageComponents/ImageComponentLayoutSchema.tsx
@@ -1,4 +1,4 @@
-import { backgroundSchema } from "../../../components/layout/v2ComponentWrapper";
+import { backgroundSchema } from "../../../components/layout/v2ComponentWrapper.schema";
 import { optimizedImageSchema } from "../../../tina/collections/shared-fields";
 import { mediaTypeField } from "../mediaType.schema";
 export const ImageComponentLayoutSchema = [
